Controller: Changes in the methods of recover data from json object.
It's very hard to update the json file, so we need to delete it, but that make the process of sync a bit slower.
So, I'm trying to think of a method of adding a relational data bank to our project.

Signed-off by: TobiasLino <tobias.lino@fatec.sp.gov.br>